# Add support for creating repository indexes in JSON format