@sabre1041 2023-07-24
This is a really good enhancement and verified that the enhancement does function correctly. However, the use of `index.yaml` for both JSON and YAML content is confusing as to the actual format of content included in this file.

It might be interesting to consider two potential files that could contain repository metadata: `index.yaml` and `index.json` (where `index.yaml` would take precedence if both files were found)