@hiddeco 2023-07-24
> It might be interesting to consider two potential files that could contain repository metadata: `index.yaml` and `index.json` (where `index.yaml` would take precedence if both files were found)

In an iteration before creating this PR, I actually had an approach like this where [`ChartRepository#DownloadIndexFile`](https://github.com/helm/helm/blob/v3.12.2/pkg/repo/chartrepo.go#L117) would loop over possible URLs (first JSON, then YAML), while breaking the loop on the first non-error response (or returning a collection of errors if none was found).

Besides this approach causing potentially loads of HTTP 404s for the server when a YAML isn't served, after looking closer at the caching logic (more precisely the various [`CacheIndexFile`](https://github.com/helm/helm/blob/f099c2e883f43f853cdffd974ad2b8f1792cee1c/pkg/helmpath/home.go#L30-L35) calls throughout the code base, and required changes to certain [error types](https://github.com/helm/helm/blob/4e7e939f19719e056c0f6e47b02068bdca6f420c/pkg/repo/index.go#L52-L53) to be factually correct). Getting this done right appeared more painful than leveraging YAML being a superset of JSON.

Another upside of the approach taken is that it's backwards compatible without having to serve (and thus store) two different file types. A Helm repository administrator can simply decide to start serving JSON, and any version of Helm will happily be able to download and work with the new index format (albeit only the Hem version which would include this change with memory consumption improvements).