@sabre1041 2023-08-02
@hiddeco @joejulian My concern with mixing JSON and YAML content is centered around two primary areas:

1. Understanding content types within `index.yaml` file. Yes YAML is a subset of JSON, but if there was an assumption that the contents included in these files were one format or the other, unexpected parsing errors could occur
2. Backwards compatibility. Do we break it by implementing such a change?