@joejulian 2023-08-02
Other way around. Yaml is a superset of json. Anything in json is 100% compatible with yaml. Yaml has features that are not in json.

So any json content in a yaml file is guaranteed compatible.