@hiddeco 2023-08-02
> YAML is a subset of JSON, but if there was an assumption that the contents included in these files were one format or the other, unexpected parsing errors could occur

The actual parser Helm uses [converts YAML to JSON](https://github.com/kubernetes-sigs/yaml/blob/b02c3522cfd003a3879afea2f58356bd02fede0f/yaml.go#L72-L81). In addition, from the [YAML 1.2 spec (2009, revised in 2021)](https://yaml.org/spec/1.2.2/#12-yaml-history):

> [...] Its primary focus was **making YAML a strict superset of JSON**. It also removed [...]

---

> Backwards compatibility. Do we break it by implementing such a change?

As noted in my [previous comment](https://github.com/helm/helm/pull/12245#issuecomment-1647442138), the exact opposite is true. As introducing an `index.json` would mean that you either:

1. Have to continue to publish both an `index.yaml` and `index.json` file as a Helm repository administrator, as otherwise older versions of Helm are unable to work with your repository. It's also worth to note this makes publishing itself more complicated, as you now have to ensure both files are in sync.
2.  Stop publishing an `index.yaml`, which would yield your Helm repository useless for older versions of Helm.