@thernstig 2023-09-11
@hiddeco @sabre1041 this is amazing! Should it not be mentioned in the docs?