@hiddeco 2023-09-11
They are partly automatically documented: https://github.com/helm/helm-www#updating-the-helm-cli-reference-docs