@mattfarina 2023-09-27
Note, while writing the release notes I discovered this doesn't fully work. I filed a bug at https://github.com/helm/helm/issues/12422.