@hiddeco 2025-06-06
It will not help if you are not in control of the published format of the repository index (which appears to be the case based on what you describe).